[squid:S1066] Merging collapsible if statements increases the code's readability

+++ app/src/main/java/dong/lan/tuyi/util/PhotoUtil.java
@@ -104,10 +104,8 @@ public class PhotoUtil {
 
 		File file = new File(dirpath, filename);
 		// 若存在即删除-默认只保留一张
-		if (isDelete) {
-			if (file.exists()) {
-				file.delete();
-			}
+		if (isDelete && file.exists()) {
+			file.delete();
 		}
 
 		if (!file.exists()) {
@@ -156,10 +154,8 @@ public class PhotoUtil {
 
 		File file = new File(dirpath, filename);
 		// 若存在即删除-默认只保留一张
-		if (isDelete) {
-			if (file.exists()) {
-				file.delete();
-			}
+		if (isDelete && file.exists()) {
+			file.delete();
 		}
 
 		if (!file.exists()) {
